+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.20.1 (unreleased)
 
+- Added `Catalog` class
 - Added `extra_columns` option to `scan_parquet` method
 
 ## 0.20.0 (2025-06-23)